# Add 3d Array and SubArray benchmarks