--- a/src/array/ArrayBenchmarks.jl
+++ b/src/array/ArrayBenchmarks.jl
@@ -20,7 +20,9 @@ const SUITE = BenchmarkGroup()
 include("sumindex.jl")
 
 s = 500
-arrays = (makearrays(Int32, s, s)..., makearrays(Float32, s, s)..., trues(s, s))
+A3d = samerand(11,11,11)
+S3d = view(A3d, 1:10, 1:10, 1:10)
+arrays = (makearrays(Int32, s, s)..., makearrays(Float32, s, s)..., trues(s, s), A3d, S3d)
 ranges = (1:10^8, 10^8:-1:1, 1.0:1e8, linspace(1,2,10^7))
 arrays_iter = map(x -> (x, string(typeof(x))), arrays)
 ranges_iter = map(x -> (x, repr(x)), ranges)
@@ -32,13 +34,17 @@ for (A, str) in (arrays_iter..., ranges_iter...)
     g["sumeach", str]            = @benchmarkable perf_sumeach($A)
     g["sumlinear", str]          = @benchmarkable perf_sumlinear($A)
     g["sumcartesian", str]       = @benchmarkable perf_sumcartesian($A)
-    g["sumcolon", str]           = @benchmarkable perf_sumcolon($A)
-    g["sumrange", str]           = @benchmarkable perf_sumrange($A)
-    g["sumlogical", str]         = @benchmarkable perf_sumlogical($A)
-    g["sumvector", str]          = @benchmarkable perf_sumvector($A)
+    if ndims(A) <= 2
+        g["sumcolon", str]       = @benchmarkable perf_sumcolon($A)
+        g["sumrange", str]       = @benchmarkable perf_sumrange($A)
+        g["sumlogical", str]     = @benchmarkable perf_sumlogical($A)
+        g["sumvector", str]      = @benchmarkable perf_sumvector($A)
+    end
 end
 g["sub2ind"] = @benchmarkable perf_sub2ind((1000,1000,1000), 1:1000, 1:1000, 1:1000)
 g["ind2sub"] = @benchmarkable perf_ind2sub((100,100,10), 1:10^5)
+g["sum", "3darray"] = @benchmarkable sum($A3d)
+g["sum", "3dsubarray"] = @benchmarkable sum($S3d)
 
 for b in values(g)
     b.params.time_tolerance = 0.40