Ref https://github.com/JuliaLang/julia/pull/18269

Master:
```julia
julia> run(BaseBenchmarks.SUITE["array"]["index"][("sumcartesian","SubArray{Float64,3,Array{Float64,3},Tuple{UnitRange{Int64},UnitRange{Int64},UnitRange{Int64}},false}")])
BenchmarkTools.Trial: 
  samples:          10000
  evals/sample:     1
  time tolerance:   40.00%
  memory tolerance: 1.00%
  memory estimate:  608.00 bytes
  allocs estimate:  23
  minimum time:     5.72 μs (0.00% GC)
  median time:      5.80 μs (0.00% GC)
  mean time:        8.08 μs (0.00% GC)
  maximum time:     22.23 ms (0.00% GC)

julia> run(BaseBenchmarks.SUITE[["array","index",("sum","3darray")]])
BenchmarkTools.Trial: 
  samples:          10000
  evals/sample:     1
  time tolerance:   40.00%
  memory tolerance: 1.00%
  memory estimate:  0.00 bytes
  allocs estimate:  0
  minimum time:     368.00 ns (0.00% GC)
  median time:      380.00 ns (0.00% GC)
  mean time:        380.73 ns (0.00% GC)
  maximum time:     9.50 μs (0.00% GC)

julia> run(BaseBenchmarks.SUITE[["array","index",("sum","3dsubarray")]])
BenchmarkTools.Trial: 
  samples:          510
  evals/sample:     1
  time tolerance:   40.00%
  memory tolerance: 1.00%
  memory estimate:  594.45 kb
  allocs estimate:  23026
  minimum time:     1.65 ms (0.00% GC)
  median time:      1.68 ms (0.00% GC)
  mean time:        1.96 ms (5.59% GC)
  maximum time:     81.40 ms (0.00% GC)
```

With 18269:
```julia
julia> run(BaseBenchmarks.SUITE["array"]["index"][("sumcartesian","SubArray{Float64,3,Array{Float64,3},Tuple{UnitRange{Int64},UnitRange{Int64},UnitRange{Int64}},false}")])
BenchmarkTools.Trial: 
  samples:          10000
  evals/sample:     1
  time tolerance:   40.00%
  memory tolerance: 1.00%
  memory estimate:  0.00 bytes
  allocs estimate:  0
  minimum time:     2.78 μs (0.00% GC)
  median time:      2.79 μs (0.00% GC)
  mean time:        3.00 μs (0.00% GC)
  maximum time:     23.83 μs (0.00% GC)

julia> run(BaseBenchmarks.SUITE[["array","index",("sum","3darray")]])
BenchmarkTools.Trial: 
  samples:          10000
  evals/sample:     1
  time tolerance:   40.00%
  memory tolerance: 1.00%
  memory estimate:  0.00 bytes
  allocs estimate:  0
  minimum time:     264.00 ns (0.00% GC)
  median time:      273.00 ns (0.00% GC)
  mean time:        276.01 ns (0.00% GC)
  maximum time:     14.48 μs (0.00% GC)

julia> run(BaseBenchmarks.SUITE[["array","index",("sum","3dsubarray")]])
BenchmarkTools.Trial: 
  samples:          10000
  evals/sample:     1
  time tolerance:   40.00%
  memory tolerance: 1.00%
  memory estimate:  112.00 bytes
  allocs estimate:  3
  minimum time:     3.79 μs (0.00% GC)
  median time:      3.85 μs (0.00% GC)
  mean time:        11.99 μs (0.00% GC)
  maximum time:     80.96 ms (0.00% GC)
```